Substantial refactoring
Breaking changes:

* Syntax of bat_sample with MCMC has changed. Added MCMCSampling.

* Renamed AHMC to HamiltonianMC, un-exported the internal undocumented HMC related types.

* Replaced localmode by marginalmode to more accurately reflect it's meaning.

* Other minor breaking changes.

+++ src/samplers/mcmc/mcmc_sampleid.jl
@@ -12,7 +12,7 @@ struct MCMCSampleID <: SampleID
     chainid::Int32
     chaincycle::Int32
     stepno::Int64
-    sampletype::Int64
+    sampletype::Int32
 end
